Move Flink Runner classes to base package and make package private
Moving them all into the same package allows making a lot of stuff
package private. This is also in preparation of getting the Flink Runner
out of the Runner.apply().
