+++ runners/flink/runner/src/main/java/org/apache/beam/runners/flink/TranslationMode.java
@@ -15,12 +15,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.flink.translation;
+package org.apache.beam.runners.flink;
 
 /**
  * The translation mode of the Beam Pipeline.
  */
-public enum TranslationMode {
+enum TranslationMode {
 
   /** Uses the batch mode of Flink. */
   BATCH,
